Master

+++ C++_Course_Work/LoginScreen.cpp
@@ -1,7 +1,6 @@
 #include "LoginScreen.h"
 #include "SHA256.h"
 #include "binary_tree.h"
-//#include "user.h"
 #include <iostream>
 #include <conio.h>
 
@@ -12,6 +11,29 @@ LoginScreen::LoginScreen()
 LoginScreen::~LoginScreen()
 {
 }
+
+//int LoginScreen::GetNameNumber(TreeNodeUser* root, std::string s)
+//{
+//	int count = 1;
+//	int* countptr = &count;
+//	return GetNameNumberPrivate(root, s, countptr);
+//}
+//
+//int LoginScreen::GetNameNumberPrivate(TreeNodeUser* node, std::string s, int* count)
+//{
+//	if (node != nullptr)
+//	{
+//		GetNameNumberPrivate(node->left, s, count);
+//		if (node->data->GetLogin() == s)
+//		{
+//			int temp = *count;
+//			return temp;
+//		}
+//		else (*count)++;
+//		GetNameNumberPrivate(node->right, s, count);
+//	}
+//}
+
 void LoginScreen::Authorization()
 {
 	int flag = 1;
@@ -47,11 +69,13 @@ void LoginScreen::Authorization()
 		// проверка на верность логина и пароля в файле
 		BinaryTreeUser* tree_user = new BinaryTreeUser();
 		tree_user->LoadUsersFromFile("../data/users.bin");
+		//std::cout << SHA256::toString(digest);
 		if (tree_user->SearchUsers(username, SHA256::toString(digest)) == 1) // 1 - admin, 2 - user, 0 - no such user
 		{
 			flag = 0;
 			system("cls");
 			std::cout << "Welcome, " << username << std::endl;
+			this->SetName(username);
 			this->AdminMenu();
 		}
 		else
@@ -61,6 +85,7 @@ void LoginScreen::Authorization()
 				flag = 0;
 				system("cls");
 				std::cout << "Welcome, " << username << std::endl;
+				this->SetName(username);
 				this->UserMenu();
 			}
 			else
@@ -75,23 +100,13 @@ void LoginScreen::AdminMenu()
 {
 	BinaryTreeUser* tree_user = new BinaryTreeUser();
 	tree_user->LoadUsersFromFile("../data/users.bin");
-	//SHA256 sha;
-	//std::string hahapassword = "Balabol";
-	//sha.update(hahapassword);
-	//hahapassword.clear();
-	//std::array<uint8_t, 32> digest = sha.digest();
-	//User* haha = new User("Balbes", SHA256::toString(digest), 0);
-	//tree_user->Insert(haha);
-	//tree_user->PrintUsers();
-	//tree_user->SaveUsersToFile("../data/users.bin");
 	Admin* admin = new Admin();
 	BinaryTreeOrder* tree_order = new BinaryTreeOrder();
 	tree_order->LoadOrdersFromFile("../data/orders.bin");
-	//tree_order->LoadOrdersFromFile(OrdersFileName);
 	int choice;
+	int number_of_current_user;
 	do 
 	{
-		//system("cls");
 		std::cout << "Menu:" << std::endl;
 		std::cout << "---------------------------|" << std::endl;
 		std::cout << "1. Add new Order           |" << std::endl;
@@ -116,12 +131,10 @@ void LoginScreen::AdminMenu()
 		std::cin >> choice;
 		while (cin.fail())
 		{
-			//in.ignore(in.rdbuf()->in_avail());
 			cin.clear();
 			while (cin.get() != '\n');
 			std::cout << "Invalid value. Try again" << std::endl;
 			cin >> choice;
-			//throw 1;
 		}
 		switch (choice) {
 		case 1:
@@ -138,15 +151,12 @@ void LoginScreen::AdminMenu()
 			std::cin >> number;
 			while (cin.fail())
 			{
-				//in.ignore(in.rdbuf()->in_avail());
 				cin.clear();
 				while (cin.get() != '\n');
 				std::cout << "Invalid value. Try again" << std::endl;
 				cin >> number;
-				//throw 1;
 			}
 			tree_order->DeleteOrderNode(&number);
-			//deleteUser(users);
 			break;
 		case 3:
 			system("cls");
@@ -158,21 +168,16 @@ void LoginScreen::AdminMenu()
 			std::cin >> number2;
 			while (cin.fail())
 			{
-				//in.ignore(in.rdbuf()->in_avail());
 				cin.clear();
 				while (cin.get() != '\n');
 				std::cout << "Invalid value. Try again" << std::endl;
 				cin >> number2;
-				//throw 1;
 			}
 			tree_order->EditOrderNode(&number2);
 			break;
 		case 4:
 			system("cls");
 			tree_order->SaveOrdersToFile("../data/orders.bin");
-			//tree_order->SaveOrdersToFile(OrdersFileName);
-			//writeUsersToFile(users, "users.txt");
-			//std::cout << "Changes saved successfully." << std::endl;
 			break;
 		case 5:
 			system("cls");
@@ -189,7 +194,6 @@ void LoginScreen::AdminMenu()
 
 				while (cin.fail())
 				{
-					//in.ignore(in.rdbuf()->in_avail());
 					cin.clear();
 					while (cin.get() != '\n');
 					std::cout << "Invalid value. Try again" << std::endl;
@@ -203,7 +207,6 @@ void LoginScreen::AdminMenu()
 					tree_order->DeleteOrderTree();
 					tree_order->LoadOrdersFromFile("../data/orders.bin");
 					choice2 = 0;
-					//tree_order->LoadOrdersFromFile(OrdersFileName);
 					break;
 				case 2:
 					system("cls");
@@ -212,10 +215,7 @@ void LoginScreen::AdminMenu()
 					break;
 				default:
 					std::cout << "Invalid choice. Please try again." << std::endl;
-					//break;
 				}
-				//users = readUsersFromFile("users.txt");
-				//std::cout << "Changes cancelled." << std::endl;
 			} while (choice2 != 0);
 			break;
 		case 6:
@@ -236,18 +236,53 @@ void LoginScreen::AdminMenu()
 			int number3;
 			std::cin.ignore(std::cin.rdbuf()->in_avail());
 			std::cin.clear();
+			number_of_current_user = tree_user->GetNameNumber(this->GetName());
 			std::cin >> number3;
 			while (cin.fail())
 			{
-				//in.ignore(in.rdbuf()->in_avail());
 				cin.clear();
 				while (cin.get() != '\n');
 				std::cout << "Invalid value. Try again" << std::endl;
 				cin >> number3;
-				//throw 1;
 			}
-			tree_user->DeleteUserNode(&number3);
-			//deleteUser(users);
+			if (number_of_current_user == number3)
+			{
+				std::cout << "Are you sure you want to delete yourself ?" << std::endl;
+				std::cout << "1. Yes" << std::endl;
+				std::cout << "2. No" << std::endl;
+				int choice5;
+				std::cin.ignore(std::cin.rdbuf()->in_avail());
+				std::cin.clear();
+
+				do
+				{
+					std::cin >> choice5;
+
+					while (cin.fail())
+					{
+						cin.clear();
+						while (cin.get() != '\n');
+						std::cout << "Invalid value. Try again" << std::endl;
+						cin >> choice5;
+					}
+					switch (choice5)
+					{
+					case 1:
+						system("cls");
+						std::cout << "Success" << std::endl;
+						tree_user->DeleteUserNode(&number3);
+						choice5 = 0;
+						break;
+					case 2:
+						system("cls");
+						std::cout << "Canceled..." << std::endl;
+						choice5 = 0;
+						break;
+					default:
+						std::cout << "Invalid choice. Please try again." << std::endl;
+					}
+				} while (choice5 != 0);
+			}
 			break;
 		case 9:
 			system("cls");
@@ -259,12 +294,10 @@ void LoginScreen::AdminMenu()
 			std::cin >> number4;
 			while (cin.fail())
 			{
-				//in.ignore(in.rdbuf()->in_avail());
 				cin.clear();
 				while (cin.get() != '\n');
 				std::cout << "Invalid value. Try again" << std::endl;
 				cin >> number4;
-				//throw 1;
 			}
 			tree_user->EditUserNode(&number4);
 			break;
@@ -287,7 +320,6 @@ void LoginScreen::AdminMenu()
 
 				while (cin.fail())
 				{
-					//in.ignore(in.rdbuf()->in_avail());
 					cin.clear();
 					while (cin.get() != '\n');
 					std::cout << "Invalid value. Try again" << std::endl;
@@ -301,7 +333,6 @@ void LoginScreen::AdminMenu()
 					tree_user->DeleteUserTree();
 					tree_user->LoadUsersFromFile("../data/users.bin");
 					choice4 = 0;
-					//tree_order->LoadOrdersFromFile(OrdersFileName);
 					break;
 				case 2:
 					system("cls");
@@ -310,10 +341,7 @@ void LoginScreen::AdminMenu()
 					break;
 				default:
 					std::cout << "Invalid choice. Please try again." << std::endl;
-					//break;
 				}
-				//users = readUsersFromFile("users.txt");
-				//std::cout << "Changes cancelled." << std::endl;
 			} while (choice4 != 0);
 			break;
 		case 12:
@@ -337,7 +365,6 @@ void LoginScreen::AdminMenu()
 				std::cin >> choice3;
 				while (cin.fail())
 				{
-					//in.ignore(in.rdbuf()->in_avail());
 					cin.clear();
 					while (cin.get() != '\n');
 					std::cout << "Invalid value. Try again" << std::endl;
@@ -350,7 +377,6 @@ void LoginScreen::AdminMenu()
 					break;
 					std::cout << "Exiting..." << std::endl;
 				case 2:
-					//system("cls");
 					choice = 1;
 					choice3 = 0;
 					break;
@@ -359,8 +385,6 @@ void LoginScreen::AdminMenu()
 					std::cout << "Invalid choice. Please try again." << std::endl;
 				}
 			} while (choice3 != 0);
-			//users = readUsersFromFile("users.txt");
-			//std::cout << "Changes cancelled." << std::endl;
 			break;
 		default:
 			std::cout << "Invalid choice. Please try again." << std::endl;
@@ -389,12 +413,10 @@ void LoginScreen::UserMenu()
 		std::cin >> choice;
 		while (cin.fail())
 		{
-			//in.ignore(in.rdbuf()->in_avail());
 			cin.clear();
 			while (cin.get() != '\n');
 			std::cout << "Invalid value. Try again" << std::endl;
 			cin >> choice;
-			//throw 1;
 		}
 		switch (choice) {
 		case 1:
@@ -430,7 +452,6 @@ void LoginScreen::UserMenu()
 				std::cin >> choice2;
 				while (cin.fail())
 				{
-					//in.ignore(in.rdbuf()->in_avail());
 					cin.clear();
 					while (cin.get() != '\n');
 					std::cout << "Invalid value. Try again" << std::endl;
